+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "express-test",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "app.js",
   "license": "MIT",
   "dependencies": {
     "express": "^4.16.3",
