+++ public/css/styles.css
@@ -1,0 +1,149 @@
+/*Home form*/
+
+body {
+    font-family: Helvetica;
+    background-color: #CEEDEA;
+}
+
+h1 {
+    text-align: center;
+}
+
+.formDiv {
+    width: 60%;
+    margin: auto;
+    text-align: center
+}
+
+.sortform {
+    display: inline;
+}
+
+.buttonDiv {
+    display: inline;
+}
+
+.small {
+    font-size: 10px;
+    color: #333;
+}
+
+.sortNameButton,
+.sortIdButton,
+.createPoke {
+    background-color: #2FA299;
+    height: 30px;
+    width: 135px;
+    font-size: 10px;
+    border: 1px solid #000;
+    border-radius: 5px;
+    margin: 2px;
+}
+
+.sortNameButton:hover,
+.sortIdButton:hover,
+.actionButtons:hover,
+.createPoke:hover {
+    opacity: 0.9;
+}
+
+.sortNameButton:active,
+.sortIdButton:active,
+.actionButtons:active,
+.createPoke:active {
+    color: #000;
+    transform: translateY(-2px);
+}
+
+.pokemonContainer {
+    width: 60%;
+    margin: 20px auto 0 auto;
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+}
+
+.pokemonDisplay {
+    height: 280px;
+    width: 180px;
+    display: flex:
+}
+
+.pokeImage {
+    height: 140px;
+    width: 140px;
+    display: block;
+    margin: auto;
+    padding: 10px;
+    background-color: #CEEDEA;
+    border-radius: 10px;
+}
+
+.pokeName {
+    text-align: center;
+    margin-top: 0px;
+    line-height: 12px;
+}
+
+.actionForm {
+    display: inline;
+}
+
+.actionButtonDiv {
+    width: 150px;
+    text-align: center;
+    margin: auto;
+}
+
+.actionButtons {
+    width: 35px;
+    height: 35px;
+    font-size: 9px;
+    background-color: #2FA299;
+    border: 1px solid #000;
+    border-radius: 50%;
+}
+
+
+/* Add & Edit Poke Form */
+
+.newFormContainer {
+    width: 100%;
+    text-align: center;
+    display;
+    flex;
+    justify-content: center;
+}
+
+.formLabel {
+    width: 50px;
+    font-size: 13px;
+    text-align: right;
+    display: inline-block;
+    margin: 20px;
+}
+
+.formInput {
+    width: 200px;
+}
+
+.submitButton {
+    width: 100px;
+    height: 30px;
+    margin-left: 200px;
+    margin-top: 30px;
+    background-color: #2FA299;
+    border: 1px solid #000;
+    border-radius: 5px;
+    padding: 10px;
+    line-height: 10px;
+}
+
+.submitButton:hover {
+    opacity: 0.9;
+}
+
+.submitButton:active {
+    color: #000;
+    transform: translateY(-2px);
+}
