--- public/home.handlebars
@@ -1,133 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <title>Welcome To Pokedex!</title>
-    <style>
-        body {
-            font-family: Helvetica;
-            background-color: #CEEDEA;
-        }
-        h1 {
-            text-align: center;
-        }
-        .formDiv {
-            width: 60%;
-            margin: auto;
-            text-align:center
-        }
-        .sortform {
-            display: inline;
-        }
-        .buttonDiv {
-            display: inline;
-        }
-        .small {
-            font-size: 10px;
-            color: #333;
-        }
-        .sortNameButton, .sortIdButton, .createPoke {
-            background-color: #2FA299;
-            height: 30px;
-            width: 135px;
-            font-size: 10px;
-            border: 1px solid #000;
-            border-radius: 5px;
-            margin: 2px;
-        }
-        .sortNameButton:hover, .sortIdButton:hover, .actionButtons:hover, .createPoke:hover {
-            opacity: 0.9;
-        }
-        .sortNameButton:active, .sortIdButton:active, .actionButtons:active, .createPoke:active {
-            color: #000;
-            transform: translateY(-2px);
-        }
-        .pokemonContainer {
-            width: 60%;
-            margin: 20px auto 0 auto;
-            display: flex;
-            flex-wrap: wrap;
-            justify-content: center;
-        }
-        .pokemonDisplay {
-            height: 280px;
-            width: 180px;
-            display: flex:
-        }
-        .pokeImage {
-            height: 140px;
-            width: 140px;
-            display: block;
-            margin: auto;
-            padding: 10px;
-            background-color: #CEEDEA;
-            border-radius: 10px;
-        }
-        .pokeName {
-            text-align: center;
-            margin-top: 0px;
-            line-height: 12px;
-        }
-        .actionForm{
-            display: inline;
-        }
-        .actionButtonDiv {
-            width: 150px;
-            text-align: center;
-            margin: auto;
-        }
-        .actionButtons {
-            width: 35px;
-            height: 35px;
-            font-size: 9px;
-            background-color: #2FA299;
-            border: 1px solid #000;
-            border-radius: 50%;
-        }
-    </style>
-</head>
-
-<body>
-    <div class="formDiv">
-        <form class="sortform" method="GET" action="/">
-            <input type="hidden" name="sortby" value="name">
-            <h1>Welcome To Pokedex!</h1>
-            <div class="buttonDiv">
-                <input class="sortNameButton" type="submit" value="Sort Pokemons By Name">
-            </div>
-        </form>
-        <form class="sortform" method="GET" action="/">
-            <input type="hidden" name="sortby" value="id">
-            <div class="buttonDiv">
-                <input class="sortIdButton" type="submit" value="Sort Pokemons By ID">
-            </div>
-        </form>
-        <form class="sortform" method="GET" action="/pokemon/new">
-            <!-- <input type="hidden" name="sortby" value="id"> -->
-            <div class="buttonDiv">
-                <input class="createPoke" type="submit" value="Create New Pokemon">
-            </div>
-        </form>
-    </div>
-
-    <div class="pokemonContainer">
-        {{#each pokeinfo}}
-        <div class="pokemonDisplay">
-            <img class="pokeImage" src="{{this.img}}">
-            <p class="pokeName"><span class="small">#{{this.num}}</span><br><br>{{this.name}}</p>
-            <div class="actionButtonDiv">
-                <form class="actionForm" method="GET" action="/pokemon/{{this.id}}/edit">
-                    <input class="actionButtons" type="submit" value="edit">
-                </form>
-                <form class="actionForm" method="POST" action="/pokemon/{{this.id}}">
-                    <input type="hidden" name="_method" value="DELETE">
-                    <input class="actionButtons" type="submit" value="del"></a>
-                </form>
-            </div>
-        </div>
-        {{/each}}
-    </div>
-</body>
-
-</html>
