+++ public/js/script.js
@@ -1,0 +1,9 @@
+let form = document.querySelector('#deletePokeForm');
+let input = document.querySelector('#inputId');
+
+let changeAction = function() {
+    let id = this.value;
+    form.action = "/pokemon/" + id;
+}
+
+input.addEventListener('change', changeAction);
