--- public/newpokeform.handlebars
@@ -1,80 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <title>Welcome To Pokedex!</title>
-    <style>
-        body {
-            font-family: Helvetica;
-            background-color: #CEEDEA;
-        }
-        h1 {
-            text-align: center;
-        }
-        .newFormContainer {
-            width: 100%;
-            text-align: center;
-            display; flex;
-            justify-content: center;
-        }
-        .formLabel {
-            width: 50px;
-            font-size: 13px;
-            text-align: right;
-            display:inline-block;
-            margin: 20px;
-        }
-        .formInput {
-            width: 200px;
-
-        }
-        .submitButton {
-            width: 100px;
-            height: 30px;
-            margin-left: 200px;
-            margin-top: 30px;
-            background-color: #2FA299;
-            border: 1px solid #000;
-            border-radius: 5px;
-            padding: 10px;
-            line-height: 10px;
-        }
-        .submitButton:hover {
-            opacity: 0.9;
-        }
-        .submitButton:active {
-            color: #000;
-            transform: translateY(-2px);
-        }
-    </style>
-</head>
-
-<body>
-    <div class="newFormContainer">
-        <form method="POST" action="/pokemon">
-            <h1>What pokemon do you want to create?</h1>
-            <div>
-                <label class="formLabel">Name</label>
-                <input class="formInput" type="text" name="name">
-            </div>
-            <div>
-                <label class="formLabel">Image</label>
-                <input class="formInput" type="text" name="img">
-            </div>
-            <div>
-                <label class="formLabel">Height</label>
-                <input class="formInput" type="text" name="height">
-            </div>
-            <div>
-                <label class="formLabel">Weight</label>
-                <input class="formInput" type="text" name="weight">
-            </div>
-            <div>
-                <input class="submitButton" type="submit" value="Submit">
-            </div>
-        </form>
-    </div>
-</body>
-
-</html>
