+++ views/editpokeform.handlebars
@@ -4,50 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Welcome To Pokedex!</title>
-    <style>
-        body {
-            font-family: Helvetica;
-            background-color: #CEEDEA;
-        }
-        h1 {
-            text-align: center;
-        }
-        .newFormContainer {
-            width: 100%;
-            text-align: center;
-            display; flex;
-            justify-content: center;
-        }
-        .formLabel {
-            width: 50px;
-            font-size: 13px;
-            text-align: right;
-            display:inline-block;
-            margin: 20px;
-        }
-        .formInput {
-            width: 200px;
-
-        }
-        .submitButton {
-            width: 100px;
-            height: 30px;
-            margin-left: 200px;
-            margin-top: 30px;
-            background-color: #2FA299;
-            border: 1px solid #000;
-            border-radius: 5px;
-            padding: 10px;
-            line-height: 10px;
-        }
-        .submitButton:hover {
-            opacity: 0.9;
-        }
-        .submitButton:active {
-            color: #000;
-            transform: translateY(-2px);
-        }
-    </style>
+    <link rel="stylesheet" href="/css/styles.css">
 </head>
 
 <body>
