+++ views/newpokeform.handlebars
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <title>Welcome To Pokedex!</title>
+    <link rel="stylesheet" href="/css/styles.css">
+</head>
+
+<body>
+    <div class="newFormContainer">
+        <form method="POST" action="/pokemon">
+            <h1>What pokemon do you want to create?</h1>
+            <div>
+                <label class="formLabel">Name</label>
+                <input class="formInput" type="text" name="name">
+            </div>
+            <div>
+                <label class="formLabel">Image</label>
+                <input class="formInput" type="text" name="img">
+            </div>
+            <div>
+                <label class="formLabel">Height</label>
+                <input class="formInput" type="text" name="height">
+            </div>
+            <div>
+                <label class="formLabel">Weight</label>
+                <input class="formInput" type="text" name="weight">
+            </div>
+            <div>
+                <input class="submitButton" type="submit" value="Submit">
+            </div>
+        </form>
+    </div>
+</body>
+
+</html>
